requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
bcrypt
jinja2
python-jose[cryptography]
websockets
alembic
psycopg2-binary